requirements.txt
celery[redis]
flower
aspose-cells-python==24.1.0
aspose-diagram-python==24.1
aspose-imaging-python-net==23.12.0
aspose-pdf==24.1.0
aspose-words==24.1.0
Aspose.Slides==24.1.0
